github-pull-request-statistics/requirements.txt
black==19.3b0
numpy==1.17.3
PyGithub==1.43.8
